+++ js/core/draft.js
@@ -460,7 +460,7 @@ function genOrderFantasy(tx, position) {
  * @return {Array.<number>} Array of salaries, in thousands of dollars/year.
  */
 function getRookieSalaries() {
-    var rookieSalaries;
+    var i, rookieSalaries;
 
     // Default for 60 picks
     rookieSalaries = [5000, 4500, 4000, 3500, 3000, 2750, 2500, 2250, 2000, 1900, 1800, 1700, 1600, 1500, 1400, 1300, 1200, 1100, 1000, 1000, 1000, 1000, 1000, 1000, 1000, 1000, 1000, 1000, 1000, 1000, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500, 500];
@@ -472,6 +472,21 @@ function getRookieSalaries() {
     while (g.numTeams * 2 < rookieSalaries.length) {
         // Remove smallest salaries
         rookieSalaries.pop();
+    }
+
+    if (g.minContract !== 500 || g.maxContract !== 20000) {
+        for (i = 0; i < rookieSalaries.length; i++) {
+            // Subtract min
+            rookieSalaries[i] -= 500;
+
+            // Scale so max will be 1/4 the max contract
+            rookieSalaries[i] *= (0.25 * g.maxContract - g.minContract) / (4500);
+
+            // Add min back
+            rookieSalaries[i] += g.minContract;
+
+            rookieSalaries[i] = Math.round(rookieSalaries[i] / 10) * 10;
+        }
     }
 
     return rookieSalaries;